[Fleet] Add missing step showing K8 instructions for managed enrollment (#130295)
* [Fleet] Add missing step showing K85 instructions for managed enrollment

* Fix checks

* Remove duplicated text and clean up commands

* Fix failing check

(cherry picked from commit 92db72b7177fbedb48b466b28caa40c496cd3520)

# Conflicts:
#	x-pack/plugins/translations/translations/fr-FR.json
#	x-pack/plugins/translations/translations/ja-JP.json
#	x-pack/plugins/translations/translations/zh-CN.json
