+++ x-pack/plugins/fleet/public/components/agent_enrollment_flyout/installation_message.tsx
@@ -5,7 +5,7 @@
  * 2.0.
  */
 import React, { useMemo } from 'react';
-import { EuiText, EuiLink } from '@elastic/eui';
+import { EuiText, EuiLink, EuiSpacer } from '@elastic/eui';
 import { FormattedMessage } from '@kbn/i18n-react';
 
 import semverMajor from 'semver/functions/major';
@@ -14,7 +14,13 @@ import semverPatch from 'semver/functions/patch';
 
 import { useKibanaVersion } from '../../hooks';
 
-export const InstallationMessage: React.FunctionComponent = () => {
+import type { K8sMode } from './types';
+
+interface Props {
+  isK8s?: K8sMode;
+}
+
+export const InstallationMessage: React.FunctionComponent<Props> = ({ isK8s }) => {
   const kibanaVersion = useKibanaVersion();
   const kibanaVersionURLString = useMemo(
     () =>
@@ -23,37 +29,47 @@ export const InstallationMessage: React.FunctionComponent = () => {
   );
 
   return (
-    <EuiText>
-      <FormattedMessage
-        id="xpack.fleet.enrollmentInstructions.installationMessage"
-        defaultMessage="Select the appropriate platform and run commands to install, enroll, and start Elastic Agent. Reuse commands to set up agents on more than one host. For aarch64, see our {downloadLink}. For additional guidance, see our {installationLink}."
-        values={{
-          downloadLink: (
-            <EuiLink
-              target="_blank"
-              external
-              href={`https://www.elastic.co/downloads/past-releases/elastic-agent-${kibanaVersionURLString}`}
-            >
-              <FormattedMessage
-                id="xpack.fleet.enrollmentInstructions.downloadLink"
-                defaultMessage="downloads page"
-              />
-            </EuiLink>
-          ),
-          installationLink: (
-            <EuiLink
-              target="_blank"
-              external
-              href="https://www.elastic.co/guide/en/fleet/current/elastic-agent-installation.html"
-            >
-              <FormattedMessage
-                id="xpack.fleet.enrollmentInstructions.installationMessage.link"
-                defaultMessage="installation docs"
-              />
-            </EuiLink>
-          ),
-        }}
-      />
-    </EuiText>
+    <>
+      <EuiText>
+        {isK8s === 'IS_KUBERNETES' ? (
+          <FormattedMessage
+            id="xpack.fleet.agentEnrollment.stepRunAgentDescriptionk8s"
+            defaultMessage="From the directory where the Kubernetes manifest is downloaded, run the apply command."
+          />
+        ) : (
+          <FormattedMessage
+            id="xpack.fleet.enrollmentInstructions.installationMessage"
+            defaultMessage="Select the appropriate platform and run commands to install, enroll, and start Elastic Agent. Reuse commands to set up agents on more than one host. For aarch64, see our {downloadLink}. For additional guidance, see our {installationLink}."
+            values={{
+              downloadLink: (
+                <EuiLink
+                  target="_blank"
+                  external
+                  href={`https://www.elastic.co/downloads/past-releases/elastic-agent-${kibanaVersionURLString}`}
+                >
+                  <FormattedMessage
+                    id="xpack.fleet.enrollmentInstructions.downloadLink"
+                    defaultMessage="downloads page"
+                  />
+                </EuiLink>
+              ),
+              installationLink: (
+                <EuiLink
+                  target="_blank"
+                  external
+                  href="https://www.elastic.co/guide/en/fleet/current/elastic-agent-installation.html"
+                >
+                  <FormattedMessage
+                    id="xpack.fleet.enrollmentInstructions.installationMessage.link"
+                    defaultMessage="installation docs"
+                  />
+                </EuiLink>
+              ),
+            }}
+          />
+        )}
+      </EuiText>
+      <EuiSpacer size="l" />
+    </>
   );
 };
