+++ x-pack/plugins/fleet/public/components/enrollment_instructions/index.tsx
@@ -6,3 +6,4 @@
  */
 
 export { ManualInstructions } from './manual';
+export { StandaloneInstructions } from './standalone';
