+++ x-pack/plugins/translations/translations/fr-FR.json
@@ -797,6 +797,11 @@
     "xpack.lens.xyVisualization.stackedPercentageBarHorizontalLabel": "H. À barres à pourcentages",
     "xpack.lens.xyVisualization.stackedPercentageBarLabel": "Vertical à barres à pourcentages",
     "xpack.lens.xyVisualization.xyLabel": "XY",
+    "coloring.dynamicColoring.customPalette.deleteButtonAriaLabel": "Supprimer",
+    "coloring.dynamicColoring.customPalette.colorStopsHelpPercentage": "Les types de valeurs en pourcentage sont relatifs à la plage complète des valeurs de données disponibles.",
+    "coloring.dynamicColoring.rangeType.label": "Type de valeur",
+    "coloring.dynamicColoring.rangeType.number": "Numéro",
+    "coloring.dynamicColoring.rangeType.percent": "Pour cent",
     "expressionXY.legend.filterForValueButtonAriaLabel": "Filtre pour la valeur",
     "expressionXY.legend.filterOptionsLegend": "{legendDataLabel}, options de filtre",
     "expressionXY.legend.filterOutValueButtonAriaLabel": "Filtrer la valeur",
@@ -9869,7 +9874,6 @@
     "xpack.fleet.agentEnrollment.stepConfigurePolicyAuthenticationTitle": "Sélectionner le jeton d'enregistrement",
     "xpack.fleet.agentEnrollment.stepDownloadAgentTitle": "Télécharger l'agent Elastic Agent sur votre hôte",
     "xpack.fleet.agentEnrollment.stepEnrollAndRunAgentTitle": "Enregistrer et démarrer l'agent Elastic Agent",
-    "xpack.fleet.agentEnrollment.stepRunAgentDescription": "Depuis le répertoire des agents, exécutez cette commande pour installer, enregistrer et démarrer un agent Elastic Agent. Vous pouvez réutiliser cette commande pour configurer des agents sur plusieurs hôtes. Cette action nécessite de disposer de privilèges d'administrateur.",
     "xpack.fleet.agentHealth.checkInTooltipText": "Dernier archivage le {lastCheckIn}",
     "xpack.fleet.agentHealth.healthyStatusText": "Sain",
     "xpack.fleet.agentHealth.inactiveStatusText": "Inactif",
