+++ x-pack/plugins/translations/translations/ja-JP.json
@@ -721,8 +721,6 @@
     "xpack.lens.paletteMetricGradient.label": "色",
     "xpack.lens.palettePicker.colorRangesLabel": "色の範囲",
     "xpack.lens.palettePicker.label": "カラーパレット",
-    "xpack.lens.paletteTableGradient.customize": "編集",
-    "xpack.lens.paletteTableGradient.label": "色",
     "xpack.lens.pie.addLayer": "ビジュアライゼーションレイヤーを追加",
     "xpack.lens.pie.arrayValues": "{label}には配列値が含まれます。可視化が想定通りに表示されない場合があります。",
     "xpack.lens.pie.donutLabel": "ドーナッツ",
@@ -946,6 +944,31 @@
     "xpack.lens.xyVisualization.stackedPercentageBarHorizontalLabel": "H.割合棒",
     "xpack.lens.xyVisualization.stackedPercentageBarLabel": "縦棒割合",
     "xpack.lens.xyVisualization.xyLabel": "XY",
+    "coloring.dynamicColoring.customPalette.addColor": "色を追加",
+    "coloring.dynamicColoring.customPalette.addColorAriaLabel": "色を追加",
+    "coloring.dynamicColoring.customPalette.deleteButtonAriaLabel": "削除",
+    "coloring.dynamicColoring.customPalette.disallowedEditMinMaxValues": "現在の構成にはカスタム値を設定できません",
+    "coloring.dynamicColoring.customPalette.distributeValues": "値を分散",
+    "coloring.dynamicColoring.customPalette.distributeValuesAriaLabel": "値を分散",
+    "coloring.dynamicColoring.customPalette.invalidMaxValue": "最大値は前の値より大きくなければなりません",
+    "coloring.dynamicColoring.customPalette.invalidValueOrColor": "1つ以上の色範囲に正しくない値または色が含まれています",
+    "coloring.dynamicColoring.customPalette.maximumStepsApplied": "ステップの最大数を適用しました",
+    "coloring.dynamicColoring.customPalette.maxValuePlaceholder": "最大値",
+    "coloring.dynamicColoring.customPalette.minValuePlaceholder": "最小値",
+    "coloring.dynamicColoring.customPalette.oneColorRange": "複数の色が必要です",
+    "coloring.dynamicColoring.customPalette.rangeAriaLabel": "範囲{index}",
+    "coloring.dynamicColoring.customPalette.reverseColors": "色を反転",
+    "coloring.dynamicColoring.customPalette.selectNewColor": "新しい色を選択",
+    "coloring.dynamicColoring.customPalette.setCustomMaxValue": "カスタム最小値を設定",
+    "coloring.dynamicColoring.customPalette.setCustomMinValue": "カスタム最大値を設定",
+    "coloring.dynamicColoring.customPalette.useAutoMaxValue": "最大データ値を使用",
+    "coloring.dynamicColoring.customPalette.useAutoMinValue": "最小データ値を使用",
+    "coloring.dynamicColoring.customPaletteAriaLabel": "色を反転",
+    "coloring.dynamicColoring.palettePicker.colorRangesLabel": "色の範囲",
+    "coloring.dynamicColoring.customPalette.colorStopsHelpPercentage": "割合値は使用可能なデータ値の全範囲に対して相対的です。",
+    "coloring.dynamicColoring.rangeType.label": "値型",
+    "coloring.dynamicColoring.rangeType.number": "数字",
+    "coloring.dynamicColoring.rangeType.percent": "割合（%）",
     "expressionXY.legend.filterForValueButtonAriaLabel": "値でフィルター",
     "expressionXY.legend.filterOptionsLegend": "{legendDataLabel}、フィルターオプション",
     "expressionXY.legend.filterOutValueButtonAriaLabel": "値を除外",
@@ -11684,7 +11707,6 @@
     "xpack.fleet.agentEnrollment.stepDownloadAgentForFleetServerTitle": "Fleetサーバーを一元化されたホストにダウンロード",
     "xpack.fleet.agentEnrollment.stepDownloadAgentTitle": "Elasticエージェントをホストにダウンロード",
     "xpack.fleet.agentEnrollment.stepEnrollAndRunAgentTitle": "Elasticエージェントを登録して実行",
-    "xpack.fleet.agentEnrollment.stepRunAgentDescription": "エージェントのディレクトリから、このコマンドを実行し、Elasticエージェントを、インストール、登録、起動します。このコマンドを再利用すると、複数のホストでエージェントを設定できます。管理者権限が必要です。",
     "xpack.fleet.agentEnrollment.stepRunAgentDescriptionk8s": "Kubernetesマニフェストがダウンロードされるディレクトリから適用コマンドを実行します。",
     "xpack.fleet.agentHealth.checkInTooltipText": "前回のチェックイン {lastCheckIn}",
     "xpack.fleet.agentHealth.healthyStatusText": "正常",
