+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -950,6 +950,31 @@
     "xpack.lens.xyVisualization.stackedPercentageBarHorizontalLabel": "水平百分比条形图",
     "xpack.lens.xyVisualization.stackedPercentageBarLabel": "垂直百分比条形图",
     "xpack.lens.xyVisualization.xyLabel": "XY",
+    "coloring.dynamicColoring.customPalette.addColor": "添加颜色",
+    "coloring.dynamicColoring.customPalette.addColorAriaLabel": "添加颜色",
+    "coloring.dynamicColoring.customPalette.deleteButtonAriaLabel": "删除",
+    "coloring.dynamicColoring.customPalette.disallowedEditMinMaxValues": "无法为当前配置设置定制值",
+    "coloring.dynamicColoring.customPalette.distributeValues": "分配值",
+    "coloring.dynamicColoring.customPalette.distributeValuesAriaLabel": "分配值",
+    "coloring.dynamicColoring.customPalette.invalidMaxValue": "最大值必须大于之前的值",
+    "coloring.dynamicColoring.customPalette.invalidValueOrColor": "至少一个颜色范围包含错误的值或颜色",
+    "coloring.dynamicColoring.customPalette.maximumStepsApplied": "您已应用最大步数",
+    "coloring.dynamicColoring.customPalette.maxValuePlaceholder": "最大值",
+    "coloring.dynamicColoring.customPalette.minValuePlaceholder": "最小值",
+    "coloring.dynamicColoring.customPalette.oneColorRange": "需要多种颜色",
+    "coloring.dynamicColoring.customPalette.rangeAriaLabel": "范围 {index}",
+    "coloring.dynamicColoring.customPalette.reverseColors": "反转颜色",
+    "coloring.dynamicColoring.customPalette.selectNewColor": "选择新颜色",
+    "coloring.dynamicColoring.customPalette.setCustomMaxValue": "设置定制最小值",
+    "coloring.dynamicColoring.customPalette.setCustomMinValue": "设置定制最大值",
+    "coloring.dynamicColoring.customPalette.useAutoMaxValue": "使用最大数据值",
+    "coloring.dynamicColoring.customPalette.useAutoMinValue": "使用最小数据值",
+    "coloring.dynamicColoring.customPaletteAriaLabel": "反转颜色",
+    "coloring.dynamicColoring.palettePicker.colorRangesLabel": "颜色范围",
+    "coloring.dynamicColoring.customPalette.colorStopsHelpPercentage": "百分比值是相对于全范围可用数据值的类型。",
+    "coloring.dynamicColoring.rangeType.label": "值类型",
+    "coloring.dynamicColoring.rangeType.number": "数字",
+    "coloring.dynamicColoring.rangeType.percent": "百分比",
     "expressionXY.legend.filterForValueButtonAriaLabel": "筛留值",
     "expressionXY.legend.filterOptionsLegend": "{legendDataLabel}, 筛选选项",
     "expressionXY.legend.filterOutValueButtonAriaLabel": "筛除值",
@@ -11703,7 +11728,6 @@
     "xpack.fleet.agentEnrollment.stepDownloadAgentForFleetServerTitle": "将 Fleet 服务器下载到集中式主机",
     "xpack.fleet.agentEnrollment.stepDownloadAgentTitle": "将 Elastic 代理下载到您的主机",
     "xpack.fleet.agentEnrollment.stepEnrollAndRunAgentTitle": "注册并启动 Elastic 代理",
-    "xpack.fleet.agentEnrollment.stepRunAgentDescription": "从代理目录运行此命令，以安装、注册并启动 Elastic 代理。您可以重复使用此命令在多个主机上设置代理。需要管理员权限。",
     "xpack.fleet.agentEnrollment.stepRunAgentDescriptionk8s": "从下载 Kubernetes 清单的目录运行应用命令。",
     "xpack.fleet.agentHealth.checkInTooltipText": "上次签入时间 {lastCheckIn}",
     "xpack.fleet.agentHealth.healthyStatusText": "运行正常",
